updated crud on courses and categories

+++ app/Http/Livewire/Admin/Categories/Categories.php
@@ -2,23 +2,24 @@
 
 namespace App\Http\Livewire\Admin\Categories;
 
-use App\Contracts\CategoryRepositoryInterface;
-use App\Models\Category;
 use Livewire\Component;
+use App\Models\Category;
 use Livewire\WithPagination;
-use phpDocumentor\Reflection\Types\Boolean;
+use App\Contracts\CategoryRepositoryInterface;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Categories extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
     public $name;
-    public $search = '';
+    public $trashed = false, $active = true, $search = '';
     public Category $category;
     
     public function render(CategoryRepositoryInterface $category)
     {
+        $this->authorize('Category_access');
         return view('livewire.admin.categories.categories', [
-            'categories' => $category->getAll($this->search)
+            'categories' => $category->getAll($this->search, $this->trashed, $this->active)
         ]);
     }
 
@@ -35,6 +36,7 @@ class Categories extends Component
 
     public function save(CategoryRepositoryInterface $categoryRepository)
     {
+       $this->authorize('Category_create');
        $categoryRepository->add($this->validate()) && 
        $this->reset('name');
        $this->emit('success', __('Created Successfully!'));
@@ -42,6 +44,7 @@ class Categories extends Component
 
     public function update(CategoryRepositoryInterface $categoryRepository)
     {
+        $this->authorize('Category_edit');
         $validName = $this->validate([ 'name' => 'required|unique:categories,name,' . $this->category->id]);
         $categoryRepository->update($this->category->id, $validName) && 
         $this->emit('success', __('Changes Saved!'));
@@ -50,6 +53,7 @@ class Categories extends Component
 
     public function delete(CategoryRepositoryInterface $categoryRepository)
     {
+        $this->authorize('Category_delete');
         $categoryRepository->remove($this->category) &&
         $this->emit('success', __('Deleted successfully!'));
         $this->reset('name');
@@ -57,7 +61,15 @@ class Categories extends Component
 
     public function toggleActive(Bool $active, CategoryRepositoryInterface $categoryRepository)
     {
+        $this->authorize('Category_edit');
        $categoryRepository->toggleActive($this->category, $active) && 
        $this->emit('success', __('Changes Saved!'));
+    }
+
+    public function restore($category, CategoryRepositoryInterface $categoryRepository)
+    {
+        $this->authorize('Category_delete');
+        $categoryRepository->restore($category) &&
+        $this->emit('success', __('Item restored!'));
     }
 }
